+++ consai_examples/CMakeLists.txt
@@ -37,10 +37,7 @@ install(PROGRAMS
     ${PROJECT_NAME}/decisions/side_back2.py
     ${PROJECT_NAME}/decisions/sub_attacker.py
     ${PROJECT_NAME}/decisions/substitute.py
-    ${PROJECT_NAME}/decisions/zone1.py
-    ${PROJECT_NAME}/decisions/zone2.py
-    ${PROJECT_NAME}/decisions/zone3.py
-    ${PROJECT_NAME}/decisions/zone4.py
+    ${PROJECT_NAME}/decisions/zone_defense.py
     DESTINATION lib/${PROJECT_NAME})
 
 install(DIRECTORY
